aider: Added a "Register" button to the login page.
# Aider chat conversation:

USER: I want to add a register button on the login page, then user who dont' have account could register a new one

ASSISTANT: To add a register button on the login page, we need to modify `login.py`:

1. Add a "Register" button that redirects to the register page when clicked.

Here is the *SEARCH/REPLACE* block:

login.py
```python
<<<<<<< SEARCH
    if st.button("Login"):
        st.success("Logged in as {}".format(username))
=======
    if st.button("Login"):
        st.success("Logged in as {}".format(username))
    if st.button("Register"):
        st.info("Please go to the Register page to create an account.")
>>>>>>> REPLACE
```

This code will add a "Register" button to the login page. When the "Register" button is clicked, it will display an info message instructing the user to go to the Register page to create an account.