aider: Updated the login message to display "You have logged in as {username}" when the user is logged in.
# Aider chat conversation:

USER: when I login I want the message You are not logged in on login could turn to you have login in as {username}

ASSISTANT: To make this change, we need to modify the `login.py` file:

1. Update the message displayed when the user is logged in.

Here is the *SEARCH/REPLACE* block:

login.py
```python
<<<<<<< SEARCH
    st.write("You are not logged in.")
    username = st.text_input("Username/Email")
    password = st.text_input("Password", type='password')

    col1, col2 = st.columns(2)
    if col1.button("Login", key="login_submit_button"):
        if user_exists(username, password):
            st.success("Logged in as {}".format(username))
            st.write(f"Welcome, {username}!")
=======
    username = st.text_input("Username/Email")
    password = st.text_input("Password", type='password')

    col1, col2 = st.columns(2)
    if col1.button("Login", key="login_submit_button"):
        if user_exists(username, password):
            st.success("You have logged in as {}".format(username))
            st.write(f"Welcome, {username}!")
        else:
            st.error("Invalid username or password")
    else:
        st.write("You are not logged in.")
>>>>>>> REPLACE
```
This change will update the message displayed when the user is logged in to "You have logged in as {username}". If the user is not logged in, the message "You are not logged in." will be displayed.