aider: Reverted changes to register.py to use a static sleep of 2 seconds after the success message.
# Aider chat conversation:

USER: sorry please return to last version

ASSISTANT: To revert the changes, we need to modify the `register.py` file to remove the dynamic sleep and return to the previous version where we had a static sleep of 2 seconds.

Here is the *SEARCH/REPLACE* block:

register.py
```python
<<<<<<< SEARCH
                st.success("User {} has been successfully registered.".format(username))
                st.session_state.register_time = time.time()
                if 'register_time' in st.session_state and time.time() - st.session_state.register_time > 2:
                    st.session_state.page = "Login"
                    st.experimental_rerun()
=======
                st.success("User {} has been successfully registered.".format(username))
                time.sleep(2)
                st.session_state.page = "Login"
                st.experimental_rerun()
>>>>>>> REPLACE
```

In the `REPLACE` block, I've replaced the dynamic sleep
^C KeyboardInterrupt

USER: sorry please return to last version

ASSISTANT: To revert the changes, we need to modify the `register.py` file to remove the dynamic sleep and return to the previous version where we had a static sleep of 2 seconds.

Here is the *SEARCH/REPLACE* block:

register.py
```python
<<<<<<< SEARCH
                st.success("User {} has been successfully registered.".format(username))
                st.session_state.register_time = time.time()
                if 'register_time' in st.session_state and time.time() - st.session_state.register_time > 2:
                    st.session_state.page = "Login"
                    st.experimental_rerun()
=======
                st.success("User {} has been successfully registered.".format(username))
                time.sleep(2)
                st.session_state.page = "Login"
                st.experimental_rerun()
>>>>>>> REPLACE
```

In the `REPLACE` block, I've replaced the dynamic sleep with a static sleep of 2 seconds after the success message.