aider: Changed the title "All the user data" to a subheader.
# Aider chat conversation:

USER: make it smaller

ASSISTANT: To clarify, could you please specify which part of the code or which file you would like to make smaller?

USER: I means the title "All the user data"

ASSISTANT: To make the title "All the user data" smaller, we can change it from a title to a subheader. This can be done by modifying the `home.py` file.

Here is the *SEARCH/REPLACE* block:

home.py
```python
<<<<<<< SEARCH
        st.title("All the user data")
=======
        st.subheader("All the user data")
>>>>>>> REPLACE
```

This change will make the title "All the user data" smaller by changing it to a subheader.