aider: Remove password from session state after successful registration.
# Aider chat conversation:

USER: after register successfully, I want to remove password, because I found that the password I enter on the register page, will remain and appear on the login page

ASSISTANT: To implement this change, we need to modify the `register.py` file. After the user is successfully registered, we need to clear the password from the session state.

Here is the *SEARCH/REPLACE* block:

register.py
```python
<<<<<<< SEARCH
                    st.success("User {} has been successfully registered.".format(username))
                    time.sleep(2)
                    st.session_state.page = "Login"
=======
                    st.success("User {} has been successfully registered.".format(username))
                    time.sleep(2)
                    st.session_state.page = "Login"
                    st.session_state.pop('password', None)
>>>>>>> REPLACE
```
This change will remove the password from the session state after the user is successfully registered.