aider: Clear all session state when registration is successful.
# Aider chat conversation:

USER: when register success I wnat to clear all the state on session

ASSISTANT: To clear all the session state when registration is successful, we need to modify the `register.py` file. We will replace the current session state pop commands with a single command to clear all session state.

Here is the *SEARCH/REPLACE* block:

register.py
```python
<<<<<<< SEARCH
                    st.success("User {} has been successfully registered.".format(username))
                    time.sleep(2)
                    st.session_state.pop('password', None)
                    st.session_state.page = "Login"
                    st.session_state.pop('password', None)
                    st.experimental_rerun()
=======
                    st.success("User {} has been successfully registered.".format(username))
                    time.sleep(2)
                    st.session_state.clear()
                    st.session_state.page = "Login"
                    st.experimental_rerun()
>>>>>>> REPLACE
```

This change will clear all session state when a user is successfully registered.